perf: move to fn perf testing into utils
